% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bmc_xml}
\alias{bmc_xml}
\title{Download full text xml of a BMC paper.}
\usage{
bmc_xml(obj = NULL, uris = NULL, dir = NULL)
}
\arguments{
  \item{obj}{(optional) An object of class bmc, from a call
  to \code{bmc_search}}

  \item{uris}{(optional) A uri to a xml file of a BMC
  paper}

  \item{dir}{(optional) A directory to save to. The file
  extension is forced to .xml, and the file name will be}

  \item{...}{Futher args passed on to httr::GET for
  debugging curl calls.}
}
\description{
Download full text xml of a BMC paper.
}
\examples{
\dontrun{
uri = 'http://www.biomedcentral.com/content/download/xml/1471-2393-14-71.xml'
uri = 'http://www.springerplus.com/content/download/xml/2193-1801-3-7.xml'
uri = 'http://www.microbiomejournal.com/content/download/xml/2049-2618-2-7.xml'
bmc_xml(uris=uri)
bmc_xml(uri, dir='~/')
bmc_xml(uri, verbose())

# from using bmc_search
out <- bmc_search(terms = 'science', limit=5)
dat <- bmc_xml(out)
length(dat)
library(plyr)
dat <- compact(dat)
length(dat)
dat
library(XML)
xpathApply(dat[[1]], "//abs", xmlValue)
saveXML(dat[[1]], file = 'myxml.xml')
}
}

