% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataOriginal}
\alias{dataOriginal}
\title{Training data for the classifier}
\format{
A data frame with 2001 rows and 260 columns:

There are 13 classes each with 20 examples (columns)
The timeseries outputs are all of length 2001
The order of the columns match the order of classes
}
\usage{
dataOriginal
}
\description{
This dataset was created using mathematical functions to create 'clean' data
for training the decision tree. It is available for information purposes
but should not be changed.
}
\keyword{datasets}
