% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ourLabels}
\alias{ourLabels}
\title{Classes for the training data}
\format{
A character vector of length 260
There are 13 classes each with 20 examples
The order of labels match the order of the columns in "dataOriginal"
}
\usage{
ourLabels
}
\description{
The classes are to be used with the training set "dataOriginal".
}
\keyword{datasets}
